refactor: Move MavRawFrameReader to BufferedMavConnection as it was simply a class encapsulating a single function
